start application before generating static

+++ bin/build-static.sh
@@ -1,9 +1,22 @@
 #!/bin/bash
 
-rm -fr static
-mkdir static
-cp -a dist static/
-cp -a images static/
-curl http://localhost:3001/ -o static/index.html
-ls static
-http-server static
+yarn start &
+yarn build
+
+ _pwd=$(pwd)
+
+if [[ $_pwd == *app-community-coin ]] ;
+then
+  rm -fr static
+  mkdir static
+  cp -a dist static/
+  cp -a images static/
+  curl http://localhost:3002/ -o static/index.html
+  ls static
+  if [[ $1 == --start-server ]] ;
+  then
+    http-server static
+  fi
+else
+  echo Run this script from the project root
+fi
